verify-test Move common *TestCase classes to avoid package conflict.
