+++ modules/apps/portal/portal-verify-test-util/src/main/java/com/liferay/portal/verify/test/util/BaseCompanySettingsVerifyProcessTestCase.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.verify.test;
+package com.liferay.portal.verify.test.util;
 
 import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.service.CompanyLocalService;
