+++ modules/apps/portal/portal-verify-test-util/src/main/resources/com/liferay/portal/verify/test/util/packageinfo
